Automatic titles for pages without one
For pages that don't have a title (single h1 heading),
generate one.

If such a title is being generated
and all other headings in the page are h2 or higher,
leave them as-is

However, if the page doesn't have a title
because it has multiple h1 headings,
push them all down one level to make room for the new title.

Resolves #9