Drop ast.Node wrapper and Pos type (#38)
Although active position tracking was nice,
given that it's only used for error messages,
that was a lot of overhead and noise to deal with
for not a lot of benefit.

Switch to simpler, on-demand position lookups.
This also allows dropping the pos.Pos type
and just deal with plain offsets as needed.