Support updating links in raw HTML
Adds support for updating relative links in `<a>` and `<img>` tags
in raw HTML in Markdown.
For HTML blocks, this was straightforward,
but inline HTML required was *quite* complicated.

Goldmark will split open-close tags in inline HTML into separate nodes:

    foo <b>bar</b> baz

Becomes 5 separate nodes:

    text("foo ")
    rawHTML("<b>")
    text("bar")
    rawHTML("</b>")
    text(" baz")

To be able to successfully parse and transform these HTML nodes,
we have to:

- traverse the document and match HTML tag pairs
- concatenate pairs (`<b></b>`), and parse and transform that HTML alone
- split the pairs back out into separate Raw HTML nodes
- update the document

The machinery to do this ended up being a fair bit of work.

Resolves #86