Add -unsafe flag to support ../ in summary
As a safety measure, we use os.DirFS to disallow referencing files
outside the current directory (or the `-C` directory).
Users can work around this by adding symlinks to the current directory.

This adds an `-unsafe` flag to drop this restriction,
and allow referencing files elsewhere in the file system.

As a small UX improvement, if we encounter these paths without the flag,
we'll let users know that the -unsafe flag exists.

Refs #111