Support summary includes
Adds support for including summary files from other summary files.
The contents of the included file will be nested under the heading.
The syntax for this borrows from Markdown images:

```
- ![Nested](other/sumamry.md)
```

This will load the specified summary file
and nest its items under the given heading;
equivalent to the following.

```
- Nested
  - [Foo](foo.md)
  - [Bar](bar.md)
```

There are restrictions on the included summary file:

- It can have only one section.
- Its section title's level will be ignored.
  The item depth will inform the heading level.

Refs #113
Resolves #4